handle string route numbers

+++ app/src/main/java/com/kieran/winnipegbusbackend/AgencySpecificClassFactory.kt
@@ -22,7 +22,7 @@ object AgencySpecificClassFactory {
     fun createRouteIdentifier(agencyId: Long, text: String): RouteIdentifier? {
         return try {
             when(agencyId){
-                2L -> WinnipegTransitRouteIdentifier(text.toInt())
+                2L -> WinnipegTransitRouteIdentifier(text)
                 else -> null
             }
         }catch (ex: Exception){
